daemon: consolidate private stateless IPC header
Rename the generic server header to stateless_ipc.h and fold the separate command, result, limit, and logging definitions into it. Keep the complete stateless wire protocol private to afpsld and libafpsl while giving daemon-private headers project-namespaced include guards.

Update daemon and test consumers to use the protocol-specific header.
