+++ daemon/stateless_internal.h
@@ -1,5 +1,5 @@
-#ifndef _STATELESS_INTERNAL_H_
-#define _STATELESS_INTERNAL_H_
+#ifndef NETATALK_CLIENT_DAEMON_STATELESS_INTERNAL_H
+#define NETATALK_CLIENT_DAEMON_STATELESS_INTERNAL_H
 
 #include <stddef.h>
