+++ daemon/stateless_ipc.h
@@ -1,12 +1,94 @@
-#ifndef _AFP_SERVER_H_
-#define _AFP_SERVER_H_
+#ifndef NETATALK_CLIENT_DAEMON_STATELESS_IPC_H
+#define NETATALK_CLIENT_DAEMON_STATELESS_IPC_H
 
 #include <limits.h>
+#include <stdint.h>
 
 #include "netatalk-client/afpsl.h"
 #include "lib/afp_protocol.h"
 
-#include "ipc.h"
+#define AFPSL_IPC_MAX_RESPONSE 16384
+
+/* Unix socket shared by afpsld and libafpsl. */
+#define AFPSL_IPC_SOCKET_PATH "/tmp/afp_sl"
+
+/* Structured logging trailer appended to stateless daemon responses. */
+#define AFPSL_IPC_LOG_MAGIC UINT32_C(0x4146504c)
+#define AFPSL_IPC_LOG_BUFFER_SIZE 4096
+
+struct afpsl_ipc_log_record {
+    int32_t level;
+    uint32_t message_len;
+};
+
+struct afpsl_ipc_log_footer {
+    uint32_t magic;
+    uint32_t log_len;
+};
+
+/* Stateless IPC command codes. */
+#define AFPSL_IPC_COMMAND_ATTACH 2
+#define AFPSL_IPC_COMMAND_DETACH 3
+#define AFPSL_IPC_COMMAND_STATUS 4
+#define AFPSL_IPC_COMMAND_EXIT 12
+#define AFPSL_IPC_COMMAND_CONNECT 14
+#define AFPSL_IPC_COMMAND_GETVOLID 16
+#define AFPSL_IPC_COMMAND_READDIR 19
+#define AFPSL_IPC_COMMAND_GETVOLS 20
+#define AFPSL_IPC_COMMAND_STAT 21
+#define AFPSL_IPC_COMMAND_OPEN 22
+#define AFPSL_IPC_COMMAND_READ 23
+#define AFPSL_IPC_COMMAND_CLOSE 24
+#define AFPSL_IPC_COMMAND_SERVERINFO 25
+#define AFPSL_IPC_COMMAND_GET_MOUNTPOINT 26
+#define AFPSL_IPC_COMMAND_WRITE 27
+#define AFPSL_IPC_COMMAND_CREAT 28
+#define AFPSL_IPC_COMMAND_CHMOD 29
+#define AFPSL_IPC_COMMAND_RENAME 30
+#define AFPSL_IPC_COMMAND_UNLINK 31
+#define AFPSL_IPC_COMMAND_TRUNCATE 32
+#define AFPSL_IPC_COMMAND_MKDIR 33
+#define AFPSL_IPC_COMMAND_RMDIR 34
+#define AFPSL_IPC_COMMAND_STATFS 35
+#define AFPSL_IPC_COMMAND_UTIME 36
+#define AFPSL_IPC_COMMAND_DISCONNECT 37
+#define AFPSL_IPC_COMMAND_CHANGEPW 38
+#define AFPSL_IPC_COMMAND_GETXATTR 39
+#define AFPSL_IPC_COMMAND_SETXATTR 40
+#define AFPSL_IPC_COMMAND_LISTXATTR 41
+#define AFPSL_IPC_COMMAND_REMOVEXATTR 42
+#define AFPSL_IPC_COMMAND_GETFINDERINFO 43
+#define AFPSL_IPC_COMMAND_SETFINDERINFO 44
+#define AFPSL_IPC_COMMAND_REMOVEFINDERINFO 45
+#define AFPSL_IPC_COMMAND_GETRESOURCEFORK 46
+#define AFPSL_IPC_COMMAND_SETRESOURCEFORK 47
+#define AFPSL_IPC_COMMAND_REMOVERESOURCEFORK 48
+#define AFPSL_IPC_COMMAND_TRUNCATERESOURCEFORK 49
+
+/* Stateless IPC result codes. */
+#define AFPSL_IPC_RESULT_OK 0
+#define AFPSL_IPC_RESULT_ERROR 1
+#define AFPSL_IPC_RESULT_TRYING 2
+#define AFPSL_IPC_RESULT_WARNING 3
+#define AFPSL_IPC_RESULT_ENOENT 4
+#define AFPSL_IPC_RESULT_NOTCONNECTED 5
+#define AFPSL_IPC_RESULT_NOTATTACHED 6
+#define AFPSL_IPC_RESULT_ALREADY_CONNECTED 7
+#define AFPSL_IPC_RESULT_ALREADY_ATTACHED 8
+#define AFPSL_IPC_RESULT_NOAUTHENT 9
+#define AFPSL_IPC_RESULT_ERROR_UNKNOWN 10
+#define AFPSL_IPC_RESULT_NOVOLUME 14
+#define AFPSL_IPC_RESULT_ALREADY_MOUNTED 15
+#define AFPSL_IPC_RESULT_VOLPASS_NEEDED 16
+#define AFPSL_IPC_RESULT_MOUNTPOINT_NOEXIST 17
+#define AFPSL_IPC_RESULT_NOSERVER 18
+#define AFPSL_IPC_RESULT_MOUNTPOINT_PERM 19
+#define AFPSL_IPC_RESULT_TIMEDOUT 20
+#define AFPSL_IPC_RESULT_DAEMON_ERROR 21
+#define AFPSL_IPC_RESULT_NOTSUPPORTED 22
+#define AFPSL_IPC_RESULT_ACCESS 23
+#define AFPSL_IPC_RESULT_EXIST 24
+#define AFPSL_IPC_RESULT_ENOTEMPTY 25
 
 struct afpsl_ipc_response_header {
     char result;
